+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Please also have a look at our
 
 ### Removed
 
+- Drop `atRuleArgs()` from the `AtRule` interface (#1141)
 - Remove `OutputFormat::get()` and `::set()` (#1108, #1110)
 - Drop special support for vendor prefixes (#1083)
 - Remove the IE hack in `Rule` (#995)